requirements.txt
python-telegram-bot==20.8
psycopg2-binary==2.9.9
pandas==2.1.3
persiantools==4.0.2
APScheduler==3.10.4